Extract email domain service
