+++ packages/worker/src/email/inbound-entitlements.workers.test.ts
@@ -5,10 +5,8 @@ import { utcDayKey } from '@kody-internal/shared/date-keys.ts'
 import { ensureUsageRollupsTestSchema } from '#worker/usage/test-schema.ts'
 import { createStableUserIdFromEmail } from '#worker/user-id.ts'
 import { handleInboundEmail } from './inbound.ts'
-import {
-	listEmailMessages,
-	maxDetailedEmailRejectionEventsPerDay,
-} from './repo.ts'
+import { listEmailMessages } from './repo.ts'
+import { maxDetailedEmailRejectionEventsPerDay } from './service.ts'
 import { createForwardableEmailMessage } from './test-fixtures.ts'
 import { ensureEmailTestSchema } from './test-schema.ts'
